@@ -1,7 +1,7 @@
 {
   "name": "@angular-redux/store",
-  "version": "6.6.0-1",
-  "description": "Angular 2 bindings for Redux",
+  "version": "7.0.0-3",
+  "description": "Angular bindings for Redux",
   "main": "./lib/src/index.js",
   "scripts": {
     "prebuild": "rimraf ./lib",
@@ -53,35 +53,35 @@
   },
   "homepage": "https://github.com/angular-redux/store#readme",
   "devDependencies": {
-    "@angular/animations": "^4.1.0",
-    "@angular/common": "^4.1.0",
-    "@angular/compiler": "^4.1.0",
-    "@angular/compiler-cli": "^4.1.0",
-    "@angular/core": "^4.1.0",
-    "@angular/http": "^4.1.0",
-    "@angular/platform-browser": "^4.1.0",
-    "@angular/platform-browser-dynamic": "^4.1.0",
-    "@angular/platform-server": "^4.1.0",
+    "@angular/animations": "^5.0.1",
+    "@angular/common": "^5.0.1",
+    "@angular/compiler": "^5.0.1",
+    "@angular/compiler-cli": "^5.0.1",
+    "@angular/core": "^5.0.1",
+    "@angular/http": "^5.0.1",
+    "@angular/platform-browser": "^5.0.1",
+    "@angular/platform-browser-dynamic": "^5.0.1",
+    "@angular/platform-server": "^5.0.1",
     "@types/jasmine": "2.5.38",
     "@types/node": "^6.0.36",
     "core-js": "^2.4.1",
     "jasmine": "^2.5.3",
     "nodemon": "^1.11.0",
     "nyc": "^10.2.0",
-    "redux": "^3.5.0",
-    "reflect-metadata": "^0.1.10",
+    "redux": "^3.7.2",
+    "reflect-metadata": "0.1.10",
     "rimraf": "^2.5.2",
-    "rxjs": "^5.0.1",
+    "rxjs": "^5.5.2",
     "symbol-observable": "^1.0.1",
     "ts-node": "^3.3.0",
     "tsconfig-paths": "^2.3.0",
-    "tslint": "^5.1.0",
+    "tslint": "^5.8.0",
     "typedoc": "^0.9.0",
-    "typescript": "^2.4.1",
+    "typescript": "2.4.2",
     "zone.js": "^0.8.18"
   },
   "peerDependencies": {
-    "@angular/core": "^2.4.0 || ^4.0.0 || ^5.0.0",
+    "@angular/core": "^5.0.0",
     "redux": "^3.5.0"
   },
   "nyc": {
