@@ -1,56 +1,66 @@
 import { Dispatch, Reducer } from 'redux';
 import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/distinctUntilChanged';
+import { map } from 'rxjs/operators/map';
+import { distinctUntilChanged } from 'rxjs/operators/distinctUntilChanged';
 
 import { getIn } from '../utils/get-in';
 import {
   PathSelector,
   Selector,
   Comparator,
-  resolveToFunctionSelector,
+  resolveToFunctionSelector
 } from './selectors';
 import { NgRedux } from './ng-redux';
-import { ObservableStore, } from './observable-store';
-import { registerFractalReducer, replaceLocalReducer } from './fractal-reducer-map';
+import { ObservableStore } from './observable-store';
+import {
+  registerFractalReducer,
+  replaceLocalReducer
+} from './fractal-reducer-map';
 
 /** @hidden */
 export class SubStore<State> implements ObservableStore<State> {
   constructor(
     private rootStore: NgRedux<any>,
     private basePath: PathSelector,
-    localReducer: Reducer<State>) {
-      registerFractalReducer(basePath, localReducer);
+    localReducer: Reducer<State>
+  ) {
+    registerFractalReducer(basePath, localReducer);
   }
 
   dispatch: Dispatch<State> = action =>
     this.rootStore.dispatch(
-      Object.assign({},
-      action,
-      { '@angular-redux::fractalkey': JSON.stringify(this.basePath) }))
+      Object.assign({}, action, {
+        '@angular-redux::fractalkey': JSON.stringify(this.basePath)
+      })
+    );
 
-  getState = (): State =>
-    getIn(this.rootStore.getState(), this.basePath)
+  getState = (): State => getIn(this.rootStore.getState(), this.basePath);
 
   configureSubStore = <SubState>(
     basePath: PathSelector,
-    localReducer: Reducer<SubState>): ObservableStore<SubState> =>
-      new SubStore<SubState>(
-        this.rootStore,
-        [ ...this.basePath, ...basePath],
-        localReducer)
+    localReducer: Reducer<SubState>
+  ): ObservableStore<SubState> =>
+    new SubStore<SubState>(
+      this.rootStore,
+      [...this.basePath, ...basePath],
+      localReducer
+    );
 
   select = <SelectedState>(
     selector?: Selector<State, SelectedState>,
-    comparator?: Comparator): Observable<SelectedState> => this.rootStore
+    comparator?: Comparator
+  ): Observable<SelectedState> =>
+    this.rootStore
       .select(this.basePath)
-      .map(resolveToFunctionSelector(selector))
-      .distinctUntilChanged(comparator)
+      .pipe(
+        map(resolveToFunctionSelector(selector)),
+        distinctUntilChanged(comparator)
+      );
 
-  subscribe = (listener: () => void): () => void => {
+  subscribe = (listener: () => void): (() => void) => {
     const subscription = this.select().subscribe(listener);
     return () => subscription.unsubscribe();
-  }
+  };
 
   replaceReducer = (nextLocalReducer: Reducer<State>) =>
     replaceLocalReducer(this.basePath, nextLocalReducer);
