@@ -1,10 +1,13 @@
 import { Reducer } from 'redux';
 import { NgRedux } from '../components/ng-redux';
 import { ObservableStore } from '../components/observable-store';
-import { Selector, PathSelector, Comparator, Transformer } from '../components/selectors';
-
-import 'rxjs/add/operator/let';
-import 'rxjs/add/operator/distinctUntilChanged';
+import {
+  Selector,
+  PathSelector,
+  Comparator,
+  Transformer
+} from '../components/selectors';
+import { distinctUntilChanged } from 'rxjs/operators/distinctUntilChanged';
 
 /**
  * Used with the `@WithSubStore` class decorator to define a SubStore (AKA a
@@ -45,7 +48,8 @@ const OPTIONS_KEY = '@angular-redux::substore::class::options';
  * This is therefore an own-property on the actual instance of the decorated class.
  */
 const INSTANCE_SUBSTORE_KEY = '@angular-redux::substore::instance::store';
-const INSTANCE_SELECTIONS_KEY = '@angular-redux::substore::instance::selections';
+const INSTANCE_SELECTIONS_KEY =
+  '@angular-redux::substore::instance::selections';
 
 /**
  * Used to detect when the base path changes - this allows components to dynamically adjust
@@ -59,16 +63,19 @@ const getClassOptions = (decoratedInstance: any): IFractalStoreOptions =>
 /** @hidden */
 export const setClassOptions = (
   decoratedClassConstructor: any,
-  options: IFractalStoreOptions): void => {
-    decoratedClassConstructor[OPTIONS_KEY] = options
-}
+  options: IFractalStoreOptions
+): void => {
+  decoratedClassConstructor[OPTIONS_KEY] = options;
+};
 
 // I want the store to be saved on the actual instance so
 // 1. different instances can have distinct substores if necessary
 // 2. the substore/selections will be marked for garbage collection when the
 //    instance is destroyed.
-const setInstanceStore = (decoratedInstance: any, store?: ObservableStore<any>) =>
-  decoratedInstance[INSTANCE_SUBSTORE_KEY] = store;
+const setInstanceStore = (
+  decoratedInstance: any,
+  store?: ObservableStore<any>
+) => (decoratedInstance[INSTANCE_SUBSTORE_KEY] = store);
 
 const getInstanceStore = (decoratedInstance: any): ObservableStore<any> =>
   decoratedInstance[INSTANCE_SUBSTORE_KEY];
@@ -77,27 +84,35 @@ const getInstanceSelectionMap = (decoratedInstance: any) => {
   const map = decoratedInstance[INSTANCE_SELECTIONS_KEY] || {};
   decoratedInstance[INSTANCE_SELECTIONS_KEY] = map;
   return map;
-}
+};
 
-const hasBasePathChanged = (decoratedInstance: any, basePath?: PathSelector): boolean =>
+const hasBasePathChanged = (
+  decoratedInstance: any,
+  basePath?: PathSelector
+): boolean =>
   decoratedInstance[INSTANCE_BASE_PATH_KEY] !== (basePath || []).toString();
 
-const setInstanceBasePath = (decoratedInstance: any, basePath?: PathSelector): void => {
+const setInstanceBasePath = (
+  decoratedInstance: any,
+  basePath?: PathSelector
+): void => {
   decoratedInstance[INSTANCE_BASE_PATH_KEY] = (basePath || []).toString();
-}
+};
 
 const clearInstanceState = (decoratedInstance: any) => {
   decoratedInstance[INSTANCE_SELECTIONS_KEY] = null;
   decoratedInstance[INSTANCE_SUBSTORE_KEY] = null;
   decoratedInstance[INSTANCE_BASE_PATH_KEY] = null;
-}
+};
 
 /**
  * Gets the store associated with a decorated instance (e.g. a
  * component or service)
  * @hidden
  */
-export const getBaseStore = (decoratedInstance: any): ObservableStore<any> | undefined => {
+export const getBaseStore = (
+  decoratedInstance: any
+): ObservableStore<any> | undefined => {
   // The root store hasn't been set up yet.
   if (!NgRedux.instance) {
     return undefined;
@@ -123,10 +138,10 @@ export const getBaseStore = (decoratedInstance: any): ObservableStore<any> | und
 
   const store = getInstanceStore(decoratedInstance);
   if (!store) {
-    setInstanceStore(decoratedInstance,
-      NgRedux.instance.configureSubStore(
-        basePath,
-        options.localReducer));
+    setInstanceStore(
+      decoratedInstance,
+      NgRedux.instance.configureSubStore(basePath, options.localReducer)
+    );
   }
 
   return getInstanceStore(decoratedInstance);
@@ -143,20 +158,23 @@ export const getInstanceSelection = <T>(
   key: string | symbol,
   selector: Selector<any, T>,
   transformer?: Transformer<any, T>,
-  comparator?: Comparator) => {
+  comparator?: Comparator
+) => {
   const store = getBaseStore(decoratedInstance);
 
   if (store) {
     const selections = getInstanceSelectionMap(decoratedInstance);
 
-    selections[key] = selections[key] || (!transformer ?
-      store.select(selector, comparator) :
-      store.select(selector)
-        .let(transformer)
-      .distinctUntilChanged(comparator));
+    selections[key] =
+      selections[key] ||
+      (!transformer
+        ? store.select(selector, comparator)
+        : store
+            .select(selector)
+            .pipe(transformer, distinctUntilChanged(comparator)));
 
     return selections[key];
   }
 
   return undefined;
-}
+};
